% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.r
\name{sN.normalizeDF}
\alias{sN.normalizeDF}
\title{Normalize data}
\usage{
sN.normalizeDF(dframe)
}
\arguments{
\item{dframe}{The dataframe to be normalized}
}
\value{
The normalized dataframe
}
\description{
Normalize all columns of a dataframe so that all values are in [0;1] and for each column the maximum
value is 1 and the minimum 0.

newx=(x-min(X))/(max(X)-min(X))
}

